+++ src/ipc/bitbucketIssueActions.ts
@@ -1,23 +1,10 @@
 import { Branch } from 'src/typings/git';
-import { BitbucketIssue, BitbucketSite } from '../bitbucket/model';
+import { BitbucketIssue } from '../bitbucket/model';
 import { Action } from './messaging';
 import { RepoData } from './prMessaging';
 
 export interface CopyBitbucketIssueLink extends Action {
     action: 'copyBitbucketIssueLink';
-}
-
-export interface CreateBitbucketIssueAction extends Action {
-    action: 'create';
-    site: BitbucketSite;
-    title: string;
-    description: string;
-    kind: string;
-    priority: string;
-}
-
-export function isCreateBitbucketIssueAction(a: Action): a is CreateBitbucketIssueAction {
-    return (<CreateBitbucketIssueAction>a).action === 'create';
 }
 
 export interface OpenBitbucketIssueAction extends Action {