+++ src/lib/analyticsApi.ts
@@ -44,5 +44,6 @@ export interface AnalyticsApi {
     fireExternalLinkEvent(source: string, linkId: string): Promise<void>;
     fireDeepLinkEvent(source: string, target: string): Promise<void>;
     fireOpenSettingsButtonEvent(source: string): Promise<void>;
+    fireExploreFeaturesButtonEvent(source: string): Promise<void>;
     firePipelineRerunEvent(site: DetailedSiteInfo, source: string): Promise<void>;
 }