+++ src/lib/ipc/models/common.ts
@@ -9,8 +9,9 @@ export enum FeedbackType {
 //These IDs uniquely identify webviews for viewScreenEvents
 export enum WebViewID {
     BitbucketIssueWebview = 'bitbucketIssueScreen',
+    CreateBitbucketIssueWebview = 'createBitbucketIssueScreen',
     ConfigWebview = 'atlascodeSettings',
-    OnboardingWebview = 'atlascodeOnboarding',
+    OnboardingWebview = 'atlascodeOnboardingScreen',
     WelcomeWebview = 'atlascodeWelcomeScreen',
     StartWork = 'startWork',
 }