+++ src/lib/ipc/toUI/startWork.ts
@@ -10,8 +10,11 @@ export enum StartWorkMessageType {
 }
 
 export type StartWorkMessage = ReducerAction<StartWorkMessageType.Init, StartWorkInitMessage>;
-
 export type StartWorkResponse = ReducerAction<StartWorkMessageType.StartWorkResponse, StartWorkResponseMessage>;
+
+export interface ComputedBranchNameMessage {
+    branchName: string;
+}
 
 export interface StartWorkIssueMessage {
     issue: MinimalIssue<DetailedSiteInfo>;
@@ -20,6 +23,8 @@ export interface StartWorkIssueMessage {
 export interface StartWorkInitMessage {
     issue: MinimalIssue<DetailedSiteInfo>;
     repoData: RepoData[];
+    customTemplate: string;
+    customPrefixes: string[];
 }
 
 export interface StartWorkResponseMessage {