+++ src/webviews/components/AtlLoader.tsx
@@ -1,11 +1,9 @@
-import React, { useContext } from 'react';
-import { ResourceContext } from './context';
+import React from 'react';
 
-export const AtlLoader: React.FunctionComponent = (props) => {
-    const scheme: string = useContext(ResourceContext);
+export const AtlLoader: React.FunctionComponent = () => {
     return (
         <div className="ac-atl-loader-container">
-            <img className="ac-atl-loader" src={`${scheme}images/atlassian-icon.svg`} />
+            <img className="ac-atl-loader" src={`images/atlassian-icon.svg`} />
             <div>Loading data...</div>
         </div>
     );